Revert "Remove unnecessary `as Long`"
This reverts commit f40a1b9d71396b59de79b01cc1dc22b8b5362653.

+++ integration/emit-default-values-json/test.ts
@@ -223,7 +223,7 @@ export const DefaultValuesTest = {
             break;
           }
 
-          message.long = longToNumber(reader.int64());
+          message.long = longToNumber(reader.int64() as Long);
           continue;
         case 5:
           if (tag !== 40) {
@@ -289,7 +289,7 @@ export const DefaultValuesTest = {
           break;
         case 14:
           if (tag === 112) {
-            message.repLong.push(longToNumber(reader.int64()));
+            message.repLong.push(longToNumber(reader.int64() as Long));
 
             continue;
           }
@@ -297,7 +297,7 @@ export const DefaultValuesTest = {
           if (tag === 114) {
             const end2 = reader.uint32() + reader.pos;
             while (reader.pos < end2) {
-              message.repLong.push(longToNumber(reader.int64()));
+              message.repLong.push(longToNumber(reader.int64() as Long));
             }
 
             continue;
@@ -361,7 +361,7 @@ export const DefaultValuesTest = {
             break;
           }
 
-          message.optLong = longToNumber(reader.int64());
+          message.optLong = longToNumber(reader.int64() as Long);
           continue;
         case 25:
           if (tag !== 200) {
