+++ integration/output-decode-only/google/protobuf/wrappers.ts
@@ -171,7 +171,7 @@ export const Int64Value = {
             break;
           }
 
-          message.value = longToNumber(reader.int64());
+          message.value = longToNumber(reader.int64() as Long);
           continue;
       }
       if ((tag & 7) === 4 || tag === 0) {
@@ -200,7 +200,7 @@ export const UInt64Value = {
             break;
           }
 
-          message.value = longToNumber(reader.uint64());
+          message.value = longToNumber(reader.uint64() as Long);
           continue;
       }
       if ((tag & 7) === 4 || tag === 0) {
