+++ integration/simple-long-bigint/google/protobuf/wrappers.ts
@@ -237,7 +237,7 @@ export const Int64Value = {
             break;
           }
 
-          message.value = longToBigint(reader.int64());
+          message.value = longToBigint(reader.int64() as Long);
           continue;
       }
       if ((tag & 7) === 4 || tag === 0) {
@@ -297,7 +297,7 @@ export const UInt64Value = {
             break;
           }
 
-          message.value = longToBigint(reader.uint64());
+          message.value = longToBigint(reader.uint64() as Long);
           continue;
       }
       if ((tag & 7) === 4 || tag === 0) {
