+++ integration/simple-long-string/google/protobuf/wrappers.ts
@@ -234,7 +234,7 @@ export const Int64Value = {
             break;
           }
 
-          message.value = longToString(reader.int64());
+          message.value = longToString(reader.int64() as Long);
           continue;
       }
       if ((tag & 7) === 4 || tag === 0) {
@@ -291,7 +291,7 @@ export const UInt64Value = {
             break;
           }
 
-          message.value = longToString(reader.uint64());
+          message.value = longToString(reader.uint64() as Long);
           continue;
       }
       if ((tag & 7) === 4 || tag === 0) {
