+++ integration/simple-unrecognized-enum/simple.ts
@@ -1762,7 +1762,7 @@ export const Numbers = {
             break;
           }
 
-          message.int64 = longToNumber(reader.int64());
+          message.int64 = longToNumber(reader.int64() as Long);
           continue;
         case 5:
           if (tag !== 40) {
@@ -1776,7 +1776,7 @@ export const Numbers = {
             break;
           }
 
-          message.uint64 = longToNumber(reader.uint64());
+          message.uint64 = longToNumber(reader.uint64() as Long);
           continue;
         case 7:
           if (tag !== 56) {
@@ -1790,7 +1790,7 @@ export const Numbers = {
             break;
           }
 
-          message.sint64 = longToNumber(reader.sint64());
+          message.sint64 = longToNumber(reader.sint64() as Long);
           continue;
         case 9:
           if (tag !== 77) {
@@ -1804,7 +1804,7 @@ export const Numbers = {
             break;
           }
 
-          message.fixed64 = longToNumber(reader.fixed64());
+          message.fixed64 = longToNumber(reader.fixed64() as Long);
           continue;
         case 11:
           if (tag !== 93) {
@@ -1818,7 +1818,7 @@ export const Numbers = {
             break;
           }
 
-          message.sfixed64 = longToNumber(reader.sfixed64());
+          message.sfixed64 = longToNumber(reader.sfixed64() as Long);
           continue;
       }
       if ((tag & 7) === 4 || tag === 0) {
