+++ integration/use-map-type/use-map-type.ts
@@ -590,14 +590,14 @@ export const Maps_Int64ToInt64Entry = {
             break;
           }
 
-          message.key = longToNumber(reader.int64());
+          message.key = longToNumber(reader.int64() as Long);
           continue;
         case 2:
           if (tag !== 16) {
             break;
           }
 
-          message.value = longToNumber(reader.int64());
+          message.value = longToNumber(reader.int64() as Long);
           continue;
       }
       if ((tag & 7) === 4 || tag === 0) {
