+++ integration/use-optionals-all/test.ts
@@ -243,7 +243,7 @@ export const OptionalsTest = {
             break;
           }
 
-          message.long = longToNumber(reader.int64());
+          message.long = longToNumber(reader.int64() as Long);
           continue;
         case 5:
           if (tag !== 40) {
@@ -309,7 +309,7 @@ export const OptionalsTest = {
           break;
         case 14:
           if (tag === 112) {
-            message.repLong!.push(longToNumber(reader.int64()));
+            message.repLong!.push(longToNumber(reader.int64() as Long));
 
             continue;
           }
@@ -317,7 +317,7 @@ export const OptionalsTest = {
           if (tag === 114) {
             const end2 = reader.uint32() + reader.pos;
             while (reader.pos < end2) {
-              message.repLong!.push(longToNumber(reader.int64()));
+              message.repLong!.push(longToNumber(reader.int64() as Long));
             }
 
             continue;
@@ -381,7 +381,7 @@ export const OptionalsTest = {
             break;
           }
 
-          message.optLong = longToNumber(reader.int64());
+          message.optLong = longToNumber(reader.int64() as Long);
           continue;
         case 25:
           if (tag !== 200) {
