+++ integration/use-optionals-no-undefined/test.ts
@@ -213,7 +213,7 @@ export const OptionalsTest = {
             break;
           }
 
-          message.long = longToNumber(reader.int64());
+          message.long = longToNumber(reader.int64() as Long);
           continue;
         case 5:
           if (tag !== 40) {
@@ -297,7 +297,7 @@ export const OptionalsTest = {
             if (message.repLong === undefined) {
               message.repLong = [];
             }
-            message.repLong!.push(longToNumber(reader.int64()));
+            message.repLong!.push(longToNumber(reader.int64() as Long));
 
             continue;
           }
@@ -308,7 +308,7 @@ export const OptionalsTest = {
             }
             const end2 = reader.uint32() + reader.pos;
             while (reader.pos < end2) {
-              message.repLong!.push(longToNumber(reader.int64()));
+              message.repLong!.push(longToNumber(reader.int64() as Long));
             }
 
             continue;
@@ -384,7 +384,7 @@ export const OptionalsTest = {
             break;
           }
 
-          message.optLong = longToNumber(reader.int64());
+          message.optLong = longToNumber(reader.int64() as Long);
           continue;
         case 25:
           if (tag !== 200) {
